+++ graphs.py
@@ -55,7 +55,12 @@ def get_scalar_plot(data, options):
 
 
 def bar_plot(data, options):
-    fig_options = ChainMap(options, GRAPHS_DEFAULT_OPTIONS["scalars"]["bar"].get_defaults())
+    xaxis_title = options.pop("xaxis_title")
+    yaxis_title = options.pop("yaxis_title")
+    fig_options = ChainMap(
+        options,
+        GRAPHS_DEFAULT_OPTIONS["scalars"]["bar"].get_defaults(exclude_non_plotly_options=True)
+    )
     try:
         fig = px.bar(
             data,
@@ -75,10 +80,14 @@ def bar_plot(data, options):
         raise PlottingError(f"Scalar plot error: {ve}")
 
     unit_axis = "x" if fig_options["orientation"] == "h" else "y"
-    axis_title = {f"{unit_axis}axis_title": add_unit_to_label(fig_options[unit_axis], data)}
+    axis_titles = {f"{unit_axis}axis_title": add_unit_to_label(fig_options[unit_axis], data)}
+    if xaxis_title:
+        axis_titles["xaxis_title"] = xaxis_title
+    if yaxis_title:
+        axis_titles["yaxis_title"] = yaxis_title
     fig.update_layout(
         template=GRAPHS_DEFAULT_TEMPLATE,
-        **axis_title,
+        **axis_titles,
         **GRAPHS_DEFAULT_LAYOUT
     )
     return fig