+++ preprocessing.py
@@ -122,7 +122,7 @@ def extract_graph_options(graph_div):
         "options": {
             item["props"]["id"].split("-")[1]: None if (value := item["props"]["value"]) == "" else value
             for item in graph_div
-            if item["type"] == "Dropdown"
+            if item["type"] in ("Dropdown", "Input") and "id" in item["props"]
         },
     }
     return options