+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 
 
 ## [Unreleased]
+### Added
+- warning information addressing "nan not in list" error
+
 ### Fixed
 - year filter initialization