+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 
 ### Changed
 - plotly template to "plotly_white"
+- graph text option is clearable
 
 ## [0.6.0] - 2021-06-24
 ### Added